fix parametros p1 y p2 calculado para sal pimienta

+++ TPs/Demo/src/main/java/ar/edu/untref/imagenes/Functions.java
@@ -418,7 +418,7 @@ public class Functions {
         return matrixResult;
     }
 
-    public int[][] applySaltAndPepper(int[][] matrix, List<int[]> pixelsSelected, int p1, int p2) {
+    public int[][] applySaltAndPepper(int[][] matrix, List<int[]> pixelsSelected, double p1, double p2) {
 
         int[][] matrixResult = matrix;
